+++ admin/tabs/activity/clienterrors/ClientErrorPanel.js
@@ -12,6 +12,7 @@ import {textInput, dateInput} from '@xh/hoist/desktop/cmp/input';
 import {toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
 import {button, exportButton, refreshButton} from '@xh/hoist/desktop/cmp/button';
 import {Icon} from '@xh/hoist/icon';
+import {LocalDate} from '@xh/hoist/utils/datetime';
 
 import {ClientErrorModel} from './ClientErrorModel';
 import {clientErrorDetail} from './ClientErrorDetail';
@@ -44,7 +45,8 @@ function renderToolbar(model) {
         renderDateInput({model, bind: 'endDate'}),
         button({
             icon: Icon.angleRight(),
-            onClick: () => model.adjustDates('add')
+            onClick: () => model.adjustDates('add'),
+            disabled: model.endDate >= LocalDate.today()
         }),
         button({
             icon: Icon.angleDoubleRight(),
@@ -63,6 +65,7 @@ function renderToolbar(model) {
 function renderDateInput(args) {
     return dateInput({
         popoverPosition: 'bottom',
+        valueType: 'localDate',
         width: 120,
         ...args
     });