+++ cmp/grid/GridModel.js
@@ -104,6 +104,7 @@ export class GridModel {
     static defaultContextMenuTokens = [
         'copy',
         'copyWithHeaders',
+        'copyCell',
         '-',
         'expandCollapseAll',
         '-',
@@ -178,6 +179,7 @@ export class GridModel {
         rowClassFn = null,
         groupSortFn,
         contextMenuFn,
+        experimental =  {},
         ...rest
     }) {
         this.treeMode = treeMode;
@@ -219,6 +221,7 @@ export class GridModel {
         this.colChooserModel = enableColChooser ? this.createChooserModel() : null;
         this.selModel = this.parseSelModel(selModel);
         this.stateModel = this.parseStateModel(stateModel);
+        this.experimental = experimental;
     }
 
     /**
@@ -492,6 +495,15 @@ export class GridModel {
         }
 
         this.columnState = columnState;
+    }
+
+    /**
+     * @param colId - id of the Column to get
+     * @returns {Column} - The Column with the given colId, or null if no Column was found with
+     *      the given colId
+     */
+    getColumn(colId) {
+        return this.findColumn(this.columns, colId);
     }
 
     /**