+++ core/appcontainer/MessageSourceModel.js
@@ -5,16 +5,15 @@
  * Copyright © 2019 Extremely Heavy Industries Inc.
  */
 
-import {defaults} from 'lodash';
-import {observable, action} from '@xh/hoist/mobx';
 import {HoistModel, managed} from '@xh/hoist/core';
-
+import {action, observable} from '@xh/hoist/mobx';
+import {isUndefined} from 'lodash';
 import {MessageModel} from './MessageModel';
 
 /**
- *  Supports displaying Modal Dialogs.
- *
- *  @private
+ * Supports managed display of modal message dialogs via `XH.message()` and friends.
+ * Not intended for direct application use. {@see XHClass#message()} and related for the public API.
+ * @private
  */
 @HoistModel
 export class MessageSourceModel {
@@ -24,24 +23,44 @@ export class MessageSourceModel {
     msgModels = [];
 
     message(config) {
+
+        // Default autoFocus on any confirm button, if no input control and developer has made no explicit request
+        const {confirmProps, cancelProps, input} = config;
+
+        if ((confirmProps && isUndefined(confirmProps.autoFocus)) &&
+            (!cancelProps || isUndefined(cancelProps.autoFocus)) &&
+            !input
+        ) {
+            confirmProps.autoFocus = true;
+        }
+
         const ret = new MessageModel(config);
         this.addModel(ret);
         return ret.result;
     }
 
     alert(config) {
-        config = defaults({}, config, {confirmText: 'OK'});
-        return this.message(config);
+        return this.message({
+            ...config,
+            confirmProps: {text: 'OK', ...config.confirmProps}
+        });
     }
 
     confirm(config) {
-        config = defaults({}, config, {confirmText: 'OK', cancelText: 'Cancel'});
-        return this.message(config);
+        return this.message({
+            ...config,
+            confirmProps: {text: 'OK', ...config.confirmProps},
+            cancelProps: {text: 'Cancel', ...config.cancelProps}
+        });
     }
 
     prompt(config) {
-        config = defaults({}, config, {confirmText: 'OK', cancelText: 'Cancel', input: {}});
-        return this.message(config);
+        return this.message({
+            ...config,
+            confirmProps: {text: 'OK', ...config.confirmProps},
+            cancelProps: {text: 'Cancel', ...config.cancelProps},
+            input: config.input || {}
+        });
     }
 
     //-----------------------------------
@@ -62,4 +81,5 @@ export class MessageSourceModel {
         this.msgModels = keepModels;
         cullModels.forEach(it => it.destroy());
     }
+
 }