+++ desktop/appcontainer/IdleDialog.js
@@ -14,11 +14,8 @@ import idleImage from './IdleDialogImage.png';
 
 
 /**
- * Default display of application suspension.
- *
- * This display can be overridden by applications.
- * @see AppSpec.idleDialogClass
- *
+ * Default dialog to display when the app has suspended itself due to inactivity.
+ * This display can be overridden by applications - {@see AppSpec.idleDialogClass}. *
  * @private
  */
 export const [IdleDialog, idleDialog] = hoistComponent(props => {
@@ -35,8 +32,12 @@ export const [IdleDialog, idleDialog] = hoistComponent(props => {
                 p('This application is sleeping due to inactivity.'),
                 p('Please click below to reload it.')
             ),
-            confirmIntent: 'primary',
-            confirmText: 'I\'m back',
+            confirmProps: {
+                text: "I'm back!",
+                intent: 'primary',
+                minimal: false,
+                autoFocus: true
+            },
             onConfirm: props.onReactivate
         },
         className: 'xh-idle-dialog'